refactor(apex-testing): consumers use runtime; drop class singleton; update mocks - W-23165788
